refactor: change ID attribute type from Long to Integer

+++ desafio02/src/main/java/com/compass/desafio02/domain/entities/Enrollment.java
@@ -4,14 +4,12 @@ import jakarta.persistence.*;
 //import jakarta.validation.constraints.NotNull;
 
 @Entity
-@Table(name = "tb_enrollment", uniqueConstraints = {
-        @UniqueConstraint(columnNames = {"student_id", "course_id"})
-})
+@Table(name = "tb_enrollment", uniqueConstraints = {@UniqueConstraint(columnNames = {"student_id", "course_id"})})
 public class Enrollment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     //@NotNull
     @ManyToOne
@@ -30,11 +28,11 @@ public class Enrollment {
         this.course = course;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
